seperated sql command and request pages into individual html pages and added tem, added temp login feature

+++ Server/server.js
@@ -7,6 +7,7 @@ var net = require('net');
 var fs = require('fs');
 var mysql = require('mysql');
 var nodemailer = require('nodemailer');
+var auth = require('C:/Users/Logan/Desktop/auth.json');
 
 
 /* session id examples
@@ -32,13 +33,18 @@ app.post('/login', function(req, res)
 var transporter = nodemailer.createTransport({
   service: 'gmail',
   auth: {
-    user: 'g',
-    pass: 'p'
+    user: auth.email,
+    pass: auth.password
   }
 });
 
-
 app.use(myParser.urlencoded({ extended: false }));
+
+// Handle logging in 
+app.post('/Modify', (request, response) => {
+	
+});
+
 
 app.post('/Results', (request, response) => {
 	//response.sendFile(path + "users.html");
@@ -100,33 +106,78 @@ function sum(value, req, res) {
 	return value;
 }
 
-function customerHash(cust, req, res)
-{
-	// for now just take ascii values. Will regret this later 
-	return cust.charCodeAt(0);
+function padDateZero(val, req, res) {
+	if (val.length != 2) {
+		return '0' + val;
+	}
+	return val.toString();
+}
+
+function getExpDate(exp, cur, req, res) {
+	if (cur == 0) {
+		var today = new Date();
+		var day = today.getDate();
+		var month = today.getMonth() + 1; // indexing starts at 0, fix that for visual representation
+		var year = today.getFullYear();
+	}
+	else {
+		var day = cur[1];
+		var month = cur[0]; // indexing starts at 0, fix that for visual representation
+		var year = cur[2];
+	}
+	// perm code 
+	if (exp == 13 || year == 9999)
+		return '12/31/9999';
+	// 1, 3, 6, 12 month
+	else {
+		// went over our year 
+		if (parseInt(month) + parseInt(exp) > 12) {
+			// day stays the same, months get added and remove 12 to get next year cycle, year gets added. 
+			return padDateZero((parseInt(month) + parseInt(exp) - 12)) + '/' + padDateZero(day) + '/' + (parseInt(year) + 1).toString();
+		}
+		// months get added normally and year does not increase 
+		return padDateZero((parseInt(month) + parseInt(exp))) + '/' + padDateZero(day)  + '/' + (year).toString();
+	}
 }
 
 
 app.post('/Created', (request, response) => {
 	//response.sendFile(path + "users.html");
 	const postBody = request.body;
-	console.log(postBody);
-	
-	// CODES MUST BE BROKEN INTO 4 SECTIONS TO KEEP FROM EXCEEDING INTEGER SIZE 
-	var codes = generate();
-	var code1 = codes[0].toString() + codes[1].toString() + codes[2].toString() + codes[3].toString();
-	var code2 = codes[4].toString() + codes[5].toString() + codes[6].toString() + codes[7].toString();
-	var code3 = codes[8].toString() + codes[9].toString() + codes[10].toString() + codes[11].toString();
-	var code4 = codes[12].toString() + codes[13].toString() + codes[14].toString() + codes[15].toString();
-	var cust = customerHash(postBody.customer);
-	
-	
-	var queryString = 'INSERT INTO customers(code1, code2, code3, code4, cust, product, expiration) VALUES (' + code1 + ', ' + code2 + ', ' + code3 + ', ' + code4 + ', ' + cust + ', ' + postBody.product + ', ' + postBody.expiration + ');';
-	con.query(queryString, (err,rows) => {
-		if(err) throw err;
-		
-		response.write("<html><body> Submitted code: " + code1 + "-" + code2 + "-" + code3 + "-" + code4 + " for " + postBody.customer + " for product " + postBody.product + "</body></html>");
-		response.end();
+	//console.log(postBody);
+	con.query('SELECT * FROM users WHERE username = \'' + postBody.username + '\' AND pass = \'' + postBody.password + '\';', (err,rows) => {
+		if (rows.length != 0) {
+			var codeString = "";
+			
+			// did user enter the values in correctly or forget any? 
+			if (!postBody.customer || !postBody.count || isNaN(postBody.count)) {
+				response.write("<html><body> Please make sure both customer name and count have a value</body></html>");
+				response.end();
+				return;
+			}
+			
+			for (var i = 0; i < parseInt(postBody.count, 10); i++) {
+				// CODES MUST BE BROKEN INTO 4 SECTIONS TO KEEP FROM EXCEEDING INTEGER SIZE 
+				var codes = generate();
+				var code1 = codes[0].toString() + codes[1].toString() + codes[2].toString() + codes[3].toString();
+				var code2 = codes[4].toString() + codes[5].toString() + codes[6].toString() + codes[7].toString();
+				var code3 = codes[8].toString() + codes[9].toString() + codes[10].toString() + codes[11].toString();
+				var code4 = codes[12].toString() + codes[13].toString() + codes[14].toString() + codes[15].toString();
+				codeString = codeString + " " + code1 + "-" + code2 + "-" + code3 + "-" + code4;
+				var exp = getExpDate(postBody.expiration, 0);
+				console.log(exp);
+				
+				var queryString = 'INSERT INTO customers(code1, code2, code3, code4, cust, product, exp, ip) VALUES (' + code1 + ', ' + code2 + ', ' + code3 + ', ' + code4 + ', \'' + postBody.customer + '\', ' + postBody.product + ',\' ' + exp + '\', 0);';
+				con.query(queryString, (err,rows) => {
+					if(err) throw err;
+					});
+			}
+			response.write("<html><body> Submitted codes: " + codeString + " for " + postBody.customer + " for product " + postBody.product + "</body></html>");
+			response.end();
+		} else {
+			response.write("<html><body>Make sure to enter your username and password. Contact admins if you need an account. </body></html>");
+			response.end();
+		}
 	});
 });
 
@@ -137,48 +188,51 @@ function splitCode(code, req, res) {
 	{
 		code = code.replace("\-", ""); 
 	}
-	console.log(code.toString());
 	return code.split('');
 }
 
 app.post('/Deleted', (request, response) => {
 	//response.sendFile(path + "users.html");
 	const postBody = request.body;
-	//console.log(postBody);
-	if (postBody.customer)
-	{
-		var cust = customerHash(postBody.customer);
-	
-		var queryString = 'DELETE FROM customers WHERE cust = ' + cust + ';';
-		con.query(queryString, (err,rows) => {
-			if(err) throw err;
+	con.query('SELECT * FROM users WHERE username = \'' + postBody.username + '\' AND pass = \'' + postBody.password + '\';', (err,rows) => {
+		if (rows.length != 0) {
+			//console.log(postBody);
+			if (postBody.customer)
+			{
+				var queryString = 'DELETE FROM customers WHERE cust = \'' + postBody.customer + '\';';
+				con.query(queryString, (err,rows) => {
+					if(err) throw err;
+					
+					response.write("<html><body> Removed all codes for " + postBody.customer + "</body></html>");
+					response.end();
+				});
+			} else {
+				//console.log(postBody.code);
+				if (postBody.code.length < 16) {
+					response.write("<html><body>Please enter a valid code delete. Formatting is a 16 digit number or a 16 digit number separated into groups of 4 by a -</body></html>");
+					response.end();
+					return;
+				}
+				var codes = splitCode(postBody.code);
+				var code1 = codes[0].toString() + codes[1].toString() + codes[2].toString() + codes[3].toString();
+				var code2 = codes[4].toString() + codes[5].toString() + codes[6].toString() + codes[7].toString();
+				var code3 = codes[8].toString() + codes[9].toString() + codes[10].toString() + codes[11].toString();
+				var code4 = codes[12].toString() + codes[13].toString() + codes[14].toString() + codes[15].toString();
+				console.log(code1 + "-" + code2 + "-" + code3 + "-" + code4);
 			
-			response.write("<html><body> Removed all codes for " + postBody.customer + "</body></html>");
+				var queryString = 'DELETE FROM customers WHERE code1 = ' + code1 + ' AND code2 = ' + code2 + ' AND code3 = ' + code3 + ' AND code4 = ' + code4 + ';';
+				con.query(queryString, (err,rows) => {
+					if(err) throw err;
+					
+					response.write("<html><body> Removed code " + code1 + "-" + code2 + "-" + code3 + "-" + code4 + "</body></html>");
+					response.end();
+				});
+			}
+		} else {
+			response.write("<html><body>Make sure to enter your username and password. Contact admins if you need an account. </body></html>");
 			response.end();
-		});
-	} else {
-		//console.log(postBody.code);
-		if (postBody.code.length < 16) {
-			response.write("<html><body>Please enter a valid code delete. Formatting is a 16 digit number or a 16 digit number separated into groups of 4 by a -</body></html>");
-			response.end();
-			return;
 		}
-		var codes = splitCode(postBody.code);
-		var code1 = codes[0].toString() + codes[1].toString() + codes[2].toString() + codes[3].toString();
-		var code2 = codes[4].toString() + codes[5].toString() + codes[6].toString() + codes[7].toString();
-		var code3 = codes[8].toString() + codes[9].toString() + codes[10].toString() + codes[11].toString();
-		var code4 = codes[12].toString() + codes[13].toString() + codes[14].toString() + codes[15].toString();
-		console.log(code1 + "-" + code2 + "-" + code3 + "-" + code4);
-	
-		var queryString = 'DELETE FROM customers WHERE code1 = ' + code1 + ' AND code2 = ' + code2 + ' AND code3 = ' + code3 + ' AND code4 = ' + code4 + ';';
-		con.query(queryString, (err,rows) => {
-			if(err) throw err;
-			
-			response.write("<html><body> Removed code " + code1 + "-" + code2 + "-" + code3 + "-" + code4 + "</body></html>");
-			response.end();
-		});
-	}
-	
+	});
 });
 
 function padZeros(code, req, res) {
@@ -217,28 +271,93 @@ function emailString(rows, req, res) {
 
 
 app.post("/Sent", (req,res) => {
-	
-	var queryString = 'SELECT code1, code2, code3, code4, product, expiration FROM customers WHERE cust = ' + customerHash(req.body.customer);
-	con.query(queryString, (err,rows) => {
-		var mailOptions = {
-			from: 'g',
-			to: req.body.email,
-			subject: 'Nesting Product Codes',
-			text: emailString(rows)
-		};
-		
-		transporter.sendMail(mailOptions, function(error, info){
-			if (error) {
-				console.log(error);
-			} else {
-				console.log('Email sent: ' + info.response);
-			}
-		});
-		console.log('Email sent to ' + req.body.email);
+	con.query('SELECT * FROM users WHERE username = \'' + req.body.username + '\' AND pass = \'' + req.body.password + '\';', (err,rows) => {
+		if (rows.length != 0) {
+			var queryString = 'SELECT code1, code2, code3, code4, product, exp FROM customers WHERE cust = \'' + req.body.customer + '\';';
+			con.query(queryString, (err,rows) => {
+				var mailOptions = {
+					from: auth.email,
+					to: req.body.email,
+					subject: 'Nesting Product Codes',
+					text: emailString(rows)
+				};
+				
+				transporter.sendMail(mailOptions, function(error, info){
+					if (error) {
+						console.log(error);
+					} else {
+						console.log('Email sent: ' + info.response);
+					}
+				});
+				console.log('Email sent to ' + req.body.email);
+			});
+			res.write("<html><body>Email Sent. Return home to continue modifying tables. </body></html>");
+			res.end();
+		}
+		else {
+			response.write("<html><body>Make sure to enter your username and password. Contact admins if you need an account. </body></html>");
+			response.end();
+		}
 	});
-	res.write("<html><body>Email Sent. Return home to continue modifying tables. </body></html>");
-	res.end();
 });
+	
+app.post('/Renewed', (request, response) => {
+	//response.sendFile(path + "users.html");
+	const postBody = request.body;
+	con.query('SELECT * FROM users WHERE username = \'' + postBody.username + '\' AND pass = \'' + postBody.password + '\';', (err,rows) => {
+		if (rows.length != 0) {
+			//console.log(postBody);
+			if (postBody.customer)
+			{
+				var queryString = 'SELECT exp FROM customers WHERE cust = \'' + postBody.customer + '\';';
+				con.query(queryString, (err,rows) => {	
+					if(err) throw err;
+					
+					for (var i = 0; i < rows.length; i++) {
+						var oldDate = rows[i].exp.split('\/');			
+						var newDate = getExpDate(postBody.expiration, oldDate);
+						var queryString2 = 'UPDATE customers SET exp=\'' + newDate + '\' WHERE cust = \'' + postBody.customer + '\';';
+						con.query(queryString2, (err,rows) => {	
+							if(err) throw err;
+						});
+					}
+					response.write("<html><body> Renewed all codes for " + postBody.customer + "</body></html>");
+					response.end();
+				});
+			} else {
+				//console.log(postBody.code);
+				if (postBody.code.length < 16) {
+					response.write("<html><body>Please enter a valid code delete. Formatting is a 16 digit number or a 16 digit number separated into groups of 4 by a -</body></html>");
+					response.end();
+					return;
+				}
+				var codes = splitCode(postBody.code);
+				var code1 = codes[0].toString() + codes[1].toString() + codes[2].toString() + codes[3].toString();
+				var code2 = codes[4].toString() + codes[5].toString() + codes[6].toString() + codes[7].toString();
+				var code3 = codes[8].toString() + codes[9].toString() + codes[10].toString() + codes[11].toString();
+				var code4 = codes[12].toString() + codes[13].toString() + codes[14].toString() + codes[15].toString();
+				
+				var queryString = 'SELECT exp FROM customers WHERE code1 = ' + code1 + ' AND code2 = ' + code2 + ' AND code3 = ' + code3 + ' AND code4 = ' + code4 + ';';
+				con.query(queryString, (err,rows) => {	
+					if(err) throw err;
+					var oldDate = rows[0].exp.split('\/');			
+					var newDate = getExpDate(postBody.expiration, oldDate);
+					var queryString2 = 'UPDATE customers SET exp=\'' + newDate + '\' WHERE code1 = ' + code1 + ' AND code2 = ' + code2 + ' AND code3 = ' + code3 + ' AND code4 = ' + code4 + ';';
+					con.query(queryString2, (err,rows) => {
+						if(err) throw err;
+						response.write("<html><body> Removed code " + code1 + "-" + code2 + "-" + code3 + "-" + code4 + "</body></html>");
+						response.end();
+					});
+				});
+			}
+		} else {
+			response.write("<html><body>Make sure to enter your username and password. Contact admins if you need an account. </body></html>");
+			response.end();
+		}
+	});
+	
+});
+
 
 
 router.use(function (req,res,next) {
@@ -247,16 +366,25 @@ router.use(function (req,res,next) {
 });
 
 router.get("/",function(req,res){
-	res.sendFile(path + "index.html");
+	res.sendFile(path + "Index.html");
 });
 
-router.get("/about",function(req,res){
-	res.sendFile(path + "about.html");
+router.get("/Renew",function(req,res){
+	res.sendFile(path + "Renew.html");
 });
 
-router.get("/contact",function(req,res){
-	res.sendFile(path + "contact.html");
+router.get("/Email",function(req,res){
+	res.sendFile(path + "Email.html");
 });
+
+router.get("/Generate",function(req,res){
+	res.sendFile(path + "Generate.html");
+});
+
+router.get("/Delete",function(req,res){
+	res.sendFile(path + "Delete.html");
+});
+
 
 app.use("/",router);
 
@@ -266,7 +394,7 @@ app.use("*",function(req,res){
 
 app.listen(3000,function(){
 	console.log("Live at Port 3000");
-	//console.log(generate());
+	//console.log(new Date().getMonth());
 });
 
 app.use(myParser.urlencoded({extended : true}));
@@ -288,10 +416,10 @@ app.get("/sendmessage", function(request, response) {
 
 
 var con = mysql.createConnection({
-	host: 'localhost',
-	user: 'root',
-	password: 'test123',
-	database : 'testing'
+	host: auth.host,
+	user: auth.user,
+	password: auth.sqlPassword,
+	database : auth.database
 });
 
 con.connect(function(err) {
